requirements.txt
pandas>=1.5.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
yfinance>=0.1.70
plotly>=5.0.0
PyYAML>=6.0
scikit-learn>=1.1.0
scipy>=1.9.0
